Beta v1.0
Added scene default greed. Translation with a mouse was a little
calibrated according to current zoom. Added transparency option to an
objeect color. Now UI axes a little transprent. Also a lot of tweeking
and debugging was done. Scene class complete cleanup. Now Display class
is a main user of the scene.